+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 0.1.1 "Milano" (03.05.2015 - ...05.2015)
+### 0.1.1 "Milano" (03.05.2015 - 04.05.2015)
 * XLSXML() into- function with colors
 * $$hashKey - remove Angular's key
 * CREATE VERTEX, CREATE EDGE